added conftest.py updated test_parking.py, test_renderer.py and README.md
