+++ README.md
@@ -22,7 +22,8 @@ https://github.com/taka-rl/RL_APS/assets/157423802/930700ff-9d21-4bcc-a0a7-bce26
     │   ├── parking.py                  # parking class
     │   ├── parking_env.py              # parking environment class
     │   └── renderer.py                 # renderer class
-    ├── tests                           # unit test
+    ├── tests                           # unit tests
+    │   ├── conftest.py                 # 
     │   ├── test_car.py                 # test for Car class
     │   ├── test_parameters.py          # test for parameters
     │   ├── test_parking.py             # test for BaseParking, PerpendicularParking and ParallelParking class
